Move AWS SDK forked part into its own package
