+++ resources/forks/doc.go
@@ -1,0 +1,3 @@
+// This package contains client forks to workaround issues with upstream dependencies.
+
+package forks
